Added Resolution optimisation settings
The resolution of the raster is optimised to be equal number of steps on the stepper motors. ie. if you select a resolution of 11 steps per millimeter, but this number of steps can't be done with exactly the same number of steps each time, the resolution is adjusted to the closest resolution that can, This might mean that a resolution of 11 will become 11.42857143 because at this resolution the stepper will move exactly 7 moves every time. this eliminates the mild banding caused the the lines not being perfectly equal distance apart.